requirements.txt
celery
djangorestframework
django
telethon==0.19.1.6
redis
requests
git+https://github.com/kunaldeo/django-livesettings3
git+https://github.com/kunaldeo/django-keyedcache3
django-cors-headers
